pyproject.toml
[project]
name = "subversion-via-focal-points-repro"
version = "1.0.0"
description = "Reproduction of 'Subversion via Focal Points' paper experiments"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.25.0",
    "openai>=1.0.0",
    "datasets>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "jsonlines>=3.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
